requirements.txt
pinecone
langchain
langchain-core
langgraph
openai
pymongo
streamlit
langsmith
requests
beautifulsoup4
python-dotenv
altair
plotly
langchain-tavily
IPython